requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
tiktoken
python-dotenv
